Added fileguard into Checker.h. Made MakTask. files with realisation of choosing necessery class and method for the main programm. These changes are in the CMakeLists.txt
