+++ Mirror/CMakeLists.txt
@@ -1,3 +1,3 @@
 project( File_work LANGUAGES CXX )
 
-add_executable( Mirror Mirror.cpp Checker.cpp temp.cpp)
+add_executable( Mirror Mirror.cpp Checker.cpp MakeTask.cpp)
