+++ Mirror/Checker.h
@@ -1,4 +1,9 @@
+#pragma once
+#ifndef Checker_HPP
+#define Checker_HPP
+
 #include <string>
 
 bool CheckFormat(const std::string& SourceFile);
 bool CheckExpansion(const std::string& SourceFile);
+#endif
